tools/img_gen/requirements.txt
openai>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
